Move event grid components
