+++ conference-app/src/wasmJsMain/resources/index.html
@@ -12,14 +12,9 @@
             width: 100%;
             overflow: hidden;
         }
-        #root {
-            height: 100%;
-            width: 100%;
-        }
     </style>
 </head>
 <body>
-    <canvas id="ComposeTarget"></canvas>
     <script src="conference-app.js"></script>
 </body>
 </html>
