+++ feature/event-grid/src/commonMain/kotlin/dev/ashdavies/playground/event/grid/EventGridCallable.kt
@@ -1,4 +1,4 @@
-package dev.ashdavies.playground.event
+package dev.ashdavies.playground.event.grid
 
 import dev.ashdavies.asg.AsgConference
 import dev.ashdavies.asg.PastConferencesCallable
