+++ feature/event-grid/src/commonMain/kotlin/dev/ashdavies/playground/event/grid/EventGridPresenter.kt
@@ -1,4 +1,4 @@
-package dev.ashdavies.playground.event
+package dev.ashdavies.playground.event.grid
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -10,6 +10,7 @@ import app.cash.sqldelight.coroutines.mapToList
 import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.runtime.presenter.Presenter
 import dev.ashdavies.event.common.PlaygroundDatabase
+import dev.ashdavies.playground.event.EventScreen
 import dev.ashdavies.sql.DatabaseFactory
 import dev.ashdavies.sql.map
 import dev.ashdavies.sql.mapAsFlow
